+++ src/compiler.c
@@ -147,7 +147,7 @@ static BOOL class_compiler(char* fname)
 
 int gARGC;
 char** gARGV;
-char* gVersion = "10.2.1";
+char* gVersion = "10.2.3";
 
 char gScriptDirPath[PATH_MAX];
 BOOL gRunningCompiler = TRUE;