+++ src/interpreter.c
@@ -2291,7 +2291,7 @@ static void compiler_final()
 
 int gARGC;
 char** gARGV;
-char* gVersion = "10.2.1";
+char* gVersion = "10.2.3";
 
 char gScriptDirPath[PATH_MAX];
 BOOL gRunningCompiler = FALSE;